BolnaBackend/requirements.txt
fastapi
uvicorn
pymysql
requests
python-dotenv